# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Get a JWT access token via given credentials.",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "description": "Login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "lomiashvili.dato5@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/refresh_token": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Get new JWT access token",
                "operationId": "b278ea66824db7f7f2acdc93dc184581",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/whoami": {
            "get": {
                "tags": [
                    "Whoami"
                ],
                "summary": "Get user details",
                "description": "This endpoint allows you to retrieve user details.",
                "operationId": "e8d8be4e3d60ddee8f69ef6900bdeacc",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/blogs": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "view blogs",
                "operationId": "81b6b233b9d9bb70ad3f41f14db041dd",
                "requestBody": {
                    "description": "<b>Sort Directions</b>: asc, desc",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "page": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "pager_limit": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "12"
                                    },
                                    "order_direction": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "desc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/blogs/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "get blog",
                "operationId": "9ac8d006089734792145676fcd6aafbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": ""
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/blogs/update/{id}": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "update blog",
                "operationId": "f409483c3f537dd11d2b072bddf4561b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": ""
                    }
                ],
                "requestBody": {
                    "description": "<b>Sort Directions</b>: asc, desc",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "New Title"
                                    },
                                    "text": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "publish_date": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2024-07-20 18:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/blogs/create": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "update blog",
                "operationId": "7b095f9d569e48fc8342fef3b5695701",
                "requestBody": {
                    "description": "<b>Order Directions</b>: asc, desc",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "New Title"
                                    },
                                    "text": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "publish_date": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2024-07-20 18:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/blogs/delete/{id}": {
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "delete blog",
                "operationId": "0d315824404f52e2e5133c15d9b1eec3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": ""
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/blogs/view/{id}": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "blog increase view counter",
                "operationId": "10355dc0c97eabdb3133f82ae2a137a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": ""
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/comments/create": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "update blog",
                "operationId": "5b10a7ec50513e0842a3f0aa6bcee257",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "blog_id": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "text": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "comment text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/comments/delete/{id}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "delete comment",
                "operationId": "7cba19d4f72223e181c0c505f63bf2fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": ""
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}